refactor: centralize content hashing
